Improve guardrails logs and UI
* Show standard model dropdown selector for guardrails
* Show an explicit Guardrail error instead of DirectResponse for
  guardrail rejections
