+++ ui/src/pages/models/ProviderConfigEditor.tsx
@@ -6,8 +6,8 @@ import {
   providerReferenceName,
   visibleProviderNames,
 } from "../../config";
+import { CloudRegionCombobox } from "../../components/CloudRegionCombobox";
 import { EnumSelector } from "../../components/EnumSelector";
-import { FreeformCombobox } from "../../components/FreeformCombobox";
 import { Field, FieldGroup, Dropdown } from "../../components/Primitives";
 import { ProviderIcon } from "../../components/ProviderIcon";
 import type { SchemaHelp } from "../../schemaHelp";
@@ -173,10 +173,10 @@ export function ProviderConfigEditor(props: {
                 )}
                 hint="Optional. If unset, Vertex uses global."
               >
-                <FreeformCombobox
+                <CloudRegionCombobox
+                  cloud="google"
                   ariaLabel="Vertex region"
                   value={props.params?.vertexRegion ?? ""}
-                  options={vertexRegions}
                   onChange={(value) =>
                     patchParams({ vertexRegion: value || null })
                   }
@@ -194,10 +194,10 @@ export function ProviderConfigEditor(props: {
                 "AWS region used for Bedrock requests.",
               )}
             >
-              <FreeformCombobox
+              <CloudRegionCombobox
+                cloud="aws"
                 ariaLabel="AWS region"
                 value={props.params?.awsRegion ?? ""}
-                options={awsRegions}
                 onChange={(value) => patchParams({ awsRegion: value || null })}
                 placeholder="us-west-2"
               />
@@ -389,89 +389,6 @@ function CustomProviderSettings(props: {
 type AwsCredentialMode = "ambient" | "static";
 type GcpCredentialMode = "ambient" | "file";
 type AzureCredentialMode = "default" | "managedIdentity" | "apiKey";
-
-const awsRegions = [
-  "us-east-1",
-  "us-east-2",
-  "us-west-1",
-  "us-west-2",
-  "af-south-1",
-  "ap-east-1",
-  "ap-south-2",
-  "ap-southeast-3",
-  "ap-southeast-5",
-  "ap-southeast-4",
-  "ap-south-1",
-  "ap-southeast-6",
-  "ap-northeast-3",
-  "ap-northeast-2",
-  "ap-southeast-1",
-  "ap-southeast-2",
-  "ap-east-2",
-  "ap-southeast-7",
-  "ap-northeast-1",
-  "ca-central-1",
-  "ca-west-1",
-  "eu-central-1",
-  "eu-west-1",
-  "eu-west-2",
-  "eu-south-1",
-  "eu-west-3",
-  "eu-south-2",
-  "eu-north-1",
-  "eu-central-2",
-  "il-central-1",
-  "mx-central-1",
-  "me-south-1",
-  "me-central-1",
-  "sa-east-1",
-];
-
-const vertexRegions = [
-  "africa-south1",
-  "asia-east1",
-  "asia-east2",
-  "asia-northeast1",
-  "asia-northeast2",
-  "asia-northeast3",
-  "asia-south1",
-  "asia-south2",
-  "asia-southeast1",
-  "asia-southeast2",
-  "asia-southeast3",
-  "australia-southeast1",
-  "australia-southeast2",
-  "europe-central2",
-  "europe-north1",
-  "europe-north2",
-  "europe-southwest1",
-  "europe-west1",
-  "europe-west10",
-  "europe-west12",
-  "europe-west2",
-  "europe-west3",
-  "europe-west4",
-  "europe-west6",
-  "europe-west8",
-  "europe-west9",
-  "me-central1",
-  "me-central2",
-  "me-west1",
-  "northamerica-northeast1",
-  "northamerica-northeast2",
-  "northamerica-south1",
-  "southamerica-east1",
-  "southamerica-west1",
-  "us-central1",
-  "us-east1",
-  "us-east4",
-  "us-east5",
-  "us-south1",
-  "us-west1",
-  "us-west2",
-  "us-west3",
-  "us-west4",
-];
 
 type ProviderAuthKey = "aws" | "gcp" | "azure";
 type ProviderAuthVariant<K extends ProviderAuthKey> = Extract<
